requirements.txt
# requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-google-genai
langchain-openai
deepgram-sdk
elevenlabs
google-cloud-speech
PyJWT
sse-starlette
